+++ src/arty/e31/device-interrupts.cpp
@@ -665,7 +665,7 @@ sifive_e31arty_interrupt_handle_unused (void)
 
   while (true)
     {
-      riscv::arch::nop ();
+      riscv::arch::wfi ();
     }
 }